refactor: create interface for session span attr populator
