+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/session/orchestrator/SessionSpanAttrPopulatorImpl.kt
@@ -1,0 +1,82 @@
+package io.embrace.android.embracesdk.internal.session.orchestrator
+
+import io.embrace.android.embracesdk.internal.arch.destination.SessionSpanWriter
+import io.embrace.android.embracesdk.internal.arch.destination.SpanAttributeData
+import io.embrace.android.embracesdk.internal.capture.metadata.MetadataService
+import io.embrace.android.embracesdk.internal.capture.startup.StartupService
+import io.embrace.android.embracesdk.internal.event.EventService
+import io.embrace.android.embracesdk.internal.logs.LogService
+import io.embrace.android.embracesdk.internal.opentelemetry.embCleanExit
+import io.embrace.android.embracesdk.internal.opentelemetry.embColdStart
+import io.embrace.android.embracesdk.internal.opentelemetry.embCrashId
+import io.embrace.android.embracesdk.internal.opentelemetry.embErrorLogCount
+import io.embrace.android.embracesdk.internal.opentelemetry.embFreeDiskBytes
+import io.embrace.android.embracesdk.internal.opentelemetry.embSdkStartupDuration
+import io.embrace.android.embracesdk.internal.opentelemetry.embSessionEndType
+import io.embrace.android.embracesdk.internal.opentelemetry.embSessionNumber
+import io.embrace.android.embracesdk.internal.opentelemetry.embSessionStartType
+import io.embrace.android.embracesdk.internal.opentelemetry.embSessionStartupDuration
+import io.embrace.android.embracesdk.internal.opentelemetry.embSessionStartupThreshold
+import io.embrace.android.embracesdk.internal.opentelemetry.embState
+import io.embrace.android.embracesdk.internal.opentelemetry.embTerminated
+import io.embrace.android.embracesdk.internal.payload.LifeEventType
+import io.embrace.android.embracesdk.internal.payload.SessionZygote
+import java.util.Locale
+
+internal class SessionSpanAttrPopulatorImpl(
+    private val sessionSpanWriter: SessionSpanWriter,
+    private val eventService: EventService,
+    private val startupService: StartupService,
+    private val logService: LogService,
+    private val metadataService: MetadataService
+) : SessionSpanAttrPopulator {
+
+    override fun populateSessionSpanStartAttrs(session: SessionZygote) {
+        with(sessionSpanWriter) {
+            addCustomAttribute(SpanAttributeData(embColdStart.name, session.isColdStart.toString()))
+            addCustomAttribute(SpanAttributeData(embSessionNumber.name, session.number.toString()))
+            addCustomAttribute(SpanAttributeData(embState.name, session.appState.name.toLowerCase(Locale.US)))
+            addCustomAttribute(SpanAttributeData(embCleanExit.name, false.toString()))
+            addCustomAttribute(SpanAttributeData(embTerminated.name, true.toString()))
+
+            session.startType.toString().toLowerCase(Locale.US).let {
+                addCustomAttribute(SpanAttributeData(embSessionStartType.name, it))
+            }
+        }
+    }
+
+    override fun populateSessionSpanEndAttrs(endType: LifeEventType?, crashId: String?, coldStart: Boolean) {
+        with(sessionSpanWriter) {
+            addCustomAttribute(SpanAttributeData(embCleanExit.name, true.toString()))
+            addCustomAttribute(SpanAttributeData(embTerminated.name, false.toString()))
+            crashId?.let {
+                addCustomAttribute(SpanAttributeData(embCrashId.name, crashId))
+            }
+            endType?.toString()?.toLowerCase(Locale.US)?.let {
+                addCustomAttribute(SpanAttributeData(embSessionEndType.name, it))
+            }
+
+            val startupInfo = when {
+                coldStart -> eventService.getStartupMomentInfo()
+                else -> null
+            }
+            startupInfo?.let { info ->
+                addCustomAttribute(
+                    SpanAttributeData(
+                        embSdkStartupDuration.name,
+                        startupService.getSdkStartupDuration(coldStart).toString()
+                    )
+                )
+                addCustomAttribute(SpanAttributeData(embSessionStartupDuration.name, info.duration.toString()))
+                addCustomAttribute(SpanAttributeData(embSessionStartupThreshold.name, info.threshold.toString()))
+            }
+
+            val logCount = logService.findErrorLogIds().size
+            addCustomAttribute(SpanAttributeData(embErrorLogCount.name, logCount.toString()))
+
+            metadataService.getDiskUsage()?.deviceDiskFree?.let { free ->
+                addCustomAttribute(SpanAttributeData(embFreeDiskBytes.name, free.toString()))
+            }
+        }
+    }
+}
