+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/FakeSessionSpanAttrPopulator.kt
@@ -1,0 +1,18 @@
+package io.embrace.android.embracesdk.fakes
+
+import io.embrace.android.embracesdk.internal.payload.LifeEventType
+import io.embrace.android.embracesdk.internal.payload.SessionZygote
+import io.embrace.android.embracesdk.internal.session.orchestrator.SessionSpanAttrPopulator
+
+internal class FakeSessionSpanAttrPopulator : SessionSpanAttrPopulator {
+
+    override fun populateSessionSpanStartAttrs(session: SessionZygote) {
+    }
+
+    override fun populateSessionSpanEndAttrs(
+        endType: LifeEventType?,
+        crashId: String?,
+        coldStart: Boolean
+    ) {
+    }
+}
