+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/injection/FakeSessionModule.kt
@@ -3,6 +3,7 @@ package io.embrace.android.embracesdk.fakes.injection
 import io.embrace.android.embracesdk.fakes.FakePayloadFactory
 import io.embrace.android.embracesdk.fakes.FakeSessionOrchestrator
 import io.embrace.android.embracesdk.fakes.FakeSessionPropertiesService
+import io.embrace.android.embracesdk.fakes.FakeSessionSpanAttrPopulator
 import io.embrace.android.embracesdk.internal.capture.session.SessionPropertiesService
 import io.embrace.android.embracesdk.internal.injection.SessionModule
 import io.embrace.android.embracesdk.internal.session.caching.PeriodicBackgroundActivityCacher
@@ -10,11 +11,13 @@ import io.embrace.android.embracesdk.internal.session.caching.PeriodicSessionCac
 import io.embrace.android.embracesdk.internal.session.message.PayloadFactory
 import io.embrace.android.embracesdk.internal.session.message.PayloadMessageCollatorImpl
 import io.embrace.android.embracesdk.internal.session.orchestrator.SessionOrchestrator
+import io.embrace.android.embracesdk.internal.session.orchestrator.SessionSpanAttrPopulator
 
 internal class FakeSessionModule(
     override val payloadFactory: PayloadFactory = FakePayloadFactory(),
     override val sessionPropertiesService: SessionPropertiesService = FakeSessionPropertiesService(),
-    override val sessionOrchestrator: SessionOrchestrator = FakeSessionOrchestrator()
+    override val sessionOrchestrator: SessionOrchestrator = FakeSessionOrchestrator(),
+    override val sessionSpanAttrPopulator: SessionSpanAttrPopulator = FakeSessionSpanAttrPopulator()
 ) : SessionModule {
 
     override val payloadMessageCollatorImpl: PayloadMessageCollatorImpl
